Decorate simple rest client for upload capabilities
