+++ example/app.js
@@ -7,6 +7,8 @@ import jsonRestClient from 'aor-json-rest-client';
 import { Delete } from 'admin-on-rest/mui';
 import frenchMessages from 'aor-language-french';
 
+import UploadCapableRestClient from './uploadCapableRestClient';
+
 import { PostList, PostCreate, PostEdit, PostShow, PostIcon } from './posts';
 import { CommentList, CommentEdit, CommentCreate, CommentIcon } from './comments';
 
@@ -18,11 +20,8 @@ const messages = {
     en: { ...englishMessages, ...customMessages.en },
 };
 
-const restClient = jsonRestClient(data, true);
-
-const delayedRestClient = (type, resource, params) => new Promise(
-    resolve => setTimeout(() => resolve(restClient(type, resource, params)), 1000),
-);
+const restClient = UploadCapableRestClient(data, true);
+const delayedRestClient = (type, resource, params) => new Promise(resolve => setTimeout(() => resolve(restClient(type, resource, params)), 1000));
 
 render(
     <Admin restClient={delayedRestClient} title="Example Admin" locale={resolveBrowserLocale()} messages={messages}>
